Move metrics module to framework
This was necessary to resolve circular imports
