+++ tests/unit/processor/test_metrics.py
@@ -4,7 +4,7 @@ from typing import List
 
 from attr import define
 
-from logprep.abc.metric import Metric
+from logprep.framework.metric import Metric
 
 
 @define(kw_only=True)
